requirements.txt
pyodbc==5.2.0
pyspark==3.5.4
python-dotenv==1.0.1
